refactor(dto): remove DtoAssembler and rollback
